mcp-flow/requirements.txt
# Common dependencies
pyyaml>=6.0
mcp>=0.1.0
fastmcp>=0.1.0
aiohttp>=3.8.0
asyncio>=3.4.3
pydantic>=2.0.0

# Kubernetes dependencies
kubernetes>=28.1.0

# Neo4j dependencies
neo4j>=5.14.0

# FAISS dependencies
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# Ollama dependencies
ollama>=0.1.0

# DuckDuckGo dependencies
duckduckgo-search>=3.9.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0
ruff>=0.1.0
mypy>=1.5.0
